Fixed bug in soft-mask loss cuda implementations where it returned int losses instead of floats

+++ layers/src/wt3dtfmloss_cuda.c
@@ -7,7 +7,7 @@
 extern THCState *state;
 
 // =============== FWD PASS ================== //
-int Weighted3DTransformLoss_forward_cuda(
+float Weighted3DTransformLoss_forward_cuda(
 			THCudaTensor *points,
 			THCudaTensor *masks,
 			THCudaTensor *tfms,
